Add optional delay to executing adjustment. Add tests.

+++ contracts/ArbSys.sol
@@ -1,5 +1,5 @@
 pragma solidity ^0.4.24;
 
 interface ArbSys {
-    function sendTxToL1(address destination, bytes calldataForL1) external payable returns(uint);
+    function sendTxToL1(address destination, bytes calldataForL1) external payable returns(uint256);
 }
